number format, update migration, add view persediaan barang

+++ database/migrations/2023_04_22_165606_create_barang_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama_barang');
             $table->integer('harga');
             $table->string('gambar');
-            $table->integer('stok');
+            $table->double('stok',8,2);
             $table->timestamps();
         });
     }
