+++ database/migrations/2023_10_07_213601_create_detail_barang_masuk_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('bm_id')->constrained('barang_masuk')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('barang_id')->constrained('barang')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('jumlah');
+            $table->double('jumlah',8,2);
             $table->timestamps();
         });
     }
